adding pagination, env file,update at seeder, edit in blade files

+++ app/Http/Resources/OrderDetailsResource.php
@@ -4,6 +4,7 @@ namespace App\Http\Resources;
 
 use Illuminate\Http\Request;
 use Illuminate\Http\Resources\Json\JsonResource;
+use Illuminate\Support\Facades\Auth;
 
 class OrderDetailsResource extends JsonResource
 {
@@ -14,15 +15,20 @@ class OrderDetailsResource extends JsonResource
      */
     public function toArray(Request $request): array
     {
+        $actions = null;
+        if (Auth::guard('web')->check()) {
+            $actions = '<a  class="change_quantity btn btn-primary btn-sm"     data-type="increase" data-order-id="' . $this->pivot->order_id . '"  data-menu-item-id="' . $this->id . '" >+</a>
+                            <a  class="change_quantity btn btn-primary btn-sm"     data-type="decreases" data-order-id="' . $this->pivot->order_id . '"  data-menu-item-id="' . $this->id . '" > - </a>
+                            <a  class="change_quantity btn btn-danger btn-sm"  data-type="delete" data-order-id="' . $this->pivot->order_id . '" data-menu-item-id="' . $this->id . '" >Delete</a>';
+        }
+
         return [
             'product_id' => $this->id,
             'name' => $this->name,
             'quantity' => $this->pivot->quantity,
             'price' => $this->price,
             'total' => $this->pivot->total,
-            'action' => '<a  class="change_quantity btn btn-primary btn-sm"     data-type="increase" data-order-id="' . $this->pivot->order_id . '"  data-menu-item-id="' . $this->id . '" >+</a>
-                            <a  class="change_quantity btn btn-primary btn-sm"     data-type="decreases" data-order-id="' . $this->pivot->order_id . '"  data-menu-item-id="' . $this->id . '" > - </a>
-                            <a  class="change_quantity btn btn-danger btn-sm"  data-type="delete" data-order-id="' . $this->pivot->order_id . '" data-menu-item-id="' . $this->id . '" >Delete</a>'
+            'action' => $actions
         ];
     }
 }
